requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
